Refactor game provider code
