+++ src/hooks/use-difficulties.ts
@@ -24,7 +24,8 @@ const DEFAULT_DIFFICULTIES: Record<PresetDifficulties, GameDifficulty> = {
 	},
 }
 
-const CUSTOM_SAVE_KEY = 'minesweeper-custom-settings'
+const CUSTOM_SETTINGS_STORAGE_KEY = 'minesweeper-custom-settings'
+const SAVED_DIFFICULTY_STORAGE_KEY = 'minesweeper-difficulty'
 
 export default function useDifficulties() {
 	const [difficulties, setDifficulties] = React.useState<Record<string, GameDifficulty>>(() => {
@@ -32,7 +33,7 @@ export default function useDifficulties() {
 			width = 10,
 			height = 10,
 			mineCount = 20,
-		] = (localStorage.getItem(CUSTOM_SAVE_KEY) ?? '10.10.20')
+		] = (localStorage.getItem(CUSTOM_SETTINGS_STORAGE_KEY) ?? '10.10.20')
 			.split('.')
 			.map(Number)
 
@@ -47,7 +48,7 @@ export default function useDifficulties() {
 
 	const setCustomDifficulty = React.useCallback(
 		function setCustomDifficulty({ width, height, mines }: { width: number, height: number, mines: number }) {
-			localStorage.setItem(CUSTOM_SAVE_KEY, `${width}.${height}.${mines}`)
+			localStorage.setItem(CUSTOM_SETTINGS_STORAGE_KEY, `${width}.${height}.${mines}`)
 
 			setDifficulties({
 				...difficulties,
@@ -60,5 +61,27 @@ export default function useDifficulties() {
 		[difficulties]
 	)
 
-	return [difficulties, setCustomDifficulty] as [typeof difficulties, typeof setCustomDifficulty]
+	const [difficulty, setDifficulty] = React.useState<keyof typeof difficulties>(
+		() => {
+			const savedDifficulty = localStorage.getItem(SAVED_DIFFICULTY_STORAGE_KEY)
+
+			if (!savedDifficulty || !Object.keys(difficulties).includes(savedDifficulty)) {
+				return Object.keys(difficulties)[0] as keyof typeof difficulties
+			}
+
+			return savedDifficulty as keyof typeof difficulties
+		}
+	)
+
+	// Save the difficulty in local storage.
+	React.useEffect(() => {
+		localStorage.setItem(SAVED_DIFFICULTY_STORAGE_KEY, difficulty)
+	}, [difficulty])
+
+	return [difficulties, setCustomDifficulty, difficulty, setDifficulty] as [
+		typeof difficulties,
+		typeof setCustomDifficulty,
+		typeof difficulty,
+		typeof setDifficulty
+	]
 }
